+++ docs/src/changelog.md
@@ -25,9 +25,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Show total duration of simulation in the log file (info), and show the current time at
   execution of each timestep (debug).
-- Support for exponential decline in horizontal conductivity in the sbm_gwf concept. This can
-  be enabled using the `exp_conductivity` flag, which exponentially reduces the `kh0` (or
-  `conductivity`) based on the value of `gwf_f` to the actual horizontal conductivity (`k`)
+- Support for exponential decline in horizontal conductivity in the sbm\_gwf concept. This
+  can be enabled using the `exp_conductivity` flag, which exponentially reduces the `kh0`
+  (or `conductivity`) based on the value of `gwf_f` to the actual horizontal conductivity
+  (`k`).
+- An optional 1D floodplain schematization for the river flow inertial model, based on
+  provided flood volumes as a function of flood depth per river cell. See also the following
+  sections: [SBM + Local inertial river and floodplain](@ref) and [River and floodplain
+  routing](@ref) for a short description, and the following section for associated [model
+  parameters](@ref local-inertial_floodplain_params).
 
 ## v0.6.2 - 2022-09-01